requirements.txt
PyHyphen
gensim
tensorflow-gpu
pandas